modified the local type inference code to indicate that inference is best-effort
In theory, this should not fail if typechecking is not going to fail later.  In practice, it does because the typechecking algorithm does more sophisticated adjustment of formal types to the calling context.  We can avoid the need to return null here if we modify the code in Application.java to provide better expected type information.